Modify address class to teleHandle class
